+++ cterasdk/asynchronous/core/files/io.py
@@ -1,7 +1,7 @@
 import logging
-from ....cio.common import encode_request_parameter, a_await_or_future
+from ....cio.common import encode_request_parameter
 from ....cio import core as fs
-from ....exceptions.io import ResourceNotFoundError, NotADirectory, ResourceExistsError
+from ....exceptions.io import ResourceNotFoundError, ResourceExistsError
 from .. import query
 
 
@@ -68,40 +68,34 @@ async def makedirs(core, path):
             logger.debug('Resource already exists: %s', path.reference.as_posix())
 
 
-async def rename(core, path, name, *, wait=False):
+async def rename(core, path, name):
     with fs.rename(path, name) as param:
-        ref = await core.v1.api.execute('', 'moveResources', param)
-        return await a_await_or_future(core, ref, wait)
+        return await core.v1.api.execute('', 'moveResources', param)
 
 
-async def remove(core, *paths, wait=False):
+async def remove(core, *paths):
     with fs.delete(*paths) as param:
-        ref = await core.v1.api.execute('', 'deleteResources', param)
-        return await a_await_or_future(core, ref, wait)
+        return await core.v1.api.execute('', 'deleteResources', param)
 
 
-async def recover(core, *paths, wait=False):
+async def recover(core, *paths):
     with fs.recover(*paths) as param:
-        ref = await core.v1.api.execute('', 'restoreResources', param)
-        return await a_await_or_future(core, ref, wait)
+        return await core.v1.api.execute('', 'restoreResources', param)
 
 
-async def copy(core, *paths, destination=None, wait=False):
-    with fs.copy(*paths, destination=destination) as param:
-        ref = await core.v1.api.execute('', 'copyResources', param)
-        return await a_await_or_future(core, ref, wait)
+async def copy(core, *paths, destination=None, resolver=None, cursor=None):
+    with fs.copy(*paths, destination=destination, resolver=resolver, cursor=cursor) as param:
+        return await core.v1.api.execute('', 'copyResources', param)
 
 
-async def move(core, *paths, destination=None, wait=False):
-    with fs.move(*paths, destination=destination) as param:
-        ref = await core.v1.api.execute('', 'moveResources', param)
-        return await a_await_or_future(core, ref, wait)
+async def move(core, *paths, destination=None, resolver=None, cursor=None):
+    with fs.move(*paths, destination=destination, resolver=resolver, cursor=cursor) as param:
+        return await core.v1.api.execute('', 'moveResources', param)
 
 
 async def ensure_directory(core, directory, suppress_error=False):
     present, resource = await metadata(core, directory, suppress_error=True)
-    if (not present or not resource.isFolder) and not suppress_error:
-        raise NotADirectory(directory.absolute)
+    fs.ensure_directory(present, resource, directory, suppress_error)
     return resource.isFolder if present else False, resource
 
 
@@ -151,7 +145,9 @@ async def _validate_destination(core, name, destination):
     is_dir, resource = await ensure_directory(core, destination, suppress_error=True)
     if not is_dir:
         is_dir, resource = await ensure_directory(core, destination.parent)
+        fs.ensure_writeable(resource, destination.parent)
         return resource.cloudFolderInfo.uid, destination.name, destination.parent
+    fs.ensure_writeable(resource, destination)
     return resource.cloudFolderInfo.uid, name, destination