+++ cterasdk/exceptions/io.py
@@ -1,7 +1,7 @@
 from .base import CTERAException
 
 
-class RemoteStorageException(CTERAException):
+class RemoteStorageError(CTERAException):
     """
     Base Exception for Remote File Storage
 
@@ -12,61 +12,78 @@ class RemoteStorageException(CTERAException):
         self.path = path
 
 
-class ResourceNotFoundError(RemoteStorageException):
+class ResourceNotFoundError(RemoteStorageError):
 
     def __init__(self, path):
         super().__init__('Remote directory not found. Please verify the path and try again.', path)
 
 
-class NotADirectory(RemoteStorageException):
+class NotADirectory(RemoteStorageError):
 
     def __init__(self, path):
         super().__init__('Target validation error: Resource exists but it is not a directory.', path)
 
 
-class ResourceExistsError(RemoteStorageException):
+class ResourceExistsError(RemoteStorageError):
 
-    def __init__(self):
-        super().__init__('Resource already exists: a file or folder with this name already exists.')
-
-
-class PathValidationError(RemoteStorageException):
-
-    def __init__(self):
-        super().__init__('Path validation failed: the specified destination path does not exist.')
+    def __init__(self, path):
+        super().__init__('Resource already exists: a file or folder with this name already exists.', path)
 
 
-class NameSyntaxError(RemoteStorageException):
+class PathValidationError(RemoteStorageError):
 
-    def __init__(self):
-        super().__init__('Invalid name: the name contains characters that are not allowed "\\ / : ? & < > \" |".')
-
-
-class ReservedNameError(RemoteStorageException):
-
-    def __init__(self):
-        super().__init__('Reserved name error: the name is reserved and cannot be used.')
+    def __init__(self, path=None, **kwargs):  # pylint: disable=unused-argument
+        super().__init__('Path validation failed: the specified destination path does not exist.', path)
 
 
-class RestrictedPathError(RemoteStorageException):
+class NameSyntaxError(RemoteStorageError):
 
-    def __init__(self):
-        super().__init__('Creating a folder in the specified location is forbidden.')
+    def __init__(self, path):
+        super().__init__('Invalid name: the name contains characters that are not allowed "\\ / : ? & < > \" |".', path)
 
 
-class RestrictedRoot(RemoteStorageException):
+class ReservedNameError(RemoteStorageError):
+
+    def __init__(self, path):
+        super().__init__('Reserved name error: the name is reserved and cannot be used.', path)
+
+
+class RestrictedPathError(RemoteStorageError):
+
+    def __init__(self, path):
+        super().__init__('Creating a folder in the specified location is forbidden.', path)
+
+
+class RestrictedRoot(RemoteStorageError):
 
     def __init__(self):
         super().__init__('Storing files to the root directory is forbidden.', '/')
 
 
-class InsufficientPermission(RemoteStorageException):
+class PermissionDenied(RemoteStorageError):
 
-    def __init__(self):
-        super().__init__('Permission denied: You must have appropriate permissions to access this resource.')
+    def __init__(self, action, path=None):
+        super().__init__('Permission denied: Inappropriate permissions to access this resource.', path)
+        self.action = action
 
 
-class UploadException(RemoteStorageException):
+class UnwriteableScope(RemoteStorageError):
+
+    def __init__(self, path, scope):
+        super().__init__("Write access denied. This target is protected and cannot be modified.", path)
+        self.scope = scope
+
+
+class FileConflict(RemoteStorageError):
+
+    def __init__(self, action, name, cursor):
+        super().__init__('Conflict: a file with the same name already exists.')
+        self.action = action
+        self.name = name
+        self.cursor = cursor
+
+
+class UploadException(RemoteStorageError):
     """
     Upload Exception