+++ tests/ut/core/admin/test_copy.py
@@ -16,8 +16,8 @@ class TestCoreFilesBrowser(base_admin.BaseCoreTest):
         src = 'cloud/Users'
         dst = 'public'
         self._init_global_admin(execute_response=expected_response)
-        actual_response = io.copy(self._global_admin, self._get_object_path(src), destination=self._get_object_path(dst), wait=False)
-        self.assertEqual(expected_response, actual_response.ref)
+        actual_response = io.copy(self._global_admin, self._get_object_path(src), destination=self._get_object_path(dst))
+        self.assertEqual(expected_response, actual_response)
         self._global_admin.api.execute.assert_called_once_with('', 'copyResources', mock.ANY)
         expected_copy_param = self._get_expected_copy_params(src, dst)
         actual_copy_param = self._global_admin.api.execute.call_args[0][2]
@@ -26,6 +26,7 @@ class TestCoreFilesBrowser(base_admin.BaseCoreTest):
     def _get_expected_copy_params(self, src, dst):
         o = Object()
         o._classname = 'ActionResourcesParam'  # pylint: disable=protected-access
+        o.startFrom = None
         src_dst_obj = Object()
         src_dst_obj._classname = 'SrcDstParam'  # pylint: disable=protected-access
         src_path = self._get_object_path(src)