+++ tests/ut/core/user/test_browser.py
@@ -11,6 +11,7 @@ class TestCoreFilesBrowser(base_admin.BaseCoreTest):
     def setUp(self):
         super().setUp()
         self.files = CloudDrive(self._global_admin)
+        self._await_or_future_mock = self.patch_call('cterasdk.core.files.browser.await_or_future')
 
     def test_versions(self):
         path = 'cloud/Users'
@@ -57,49 +58,54 @@ class TestCoreFilesBrowser(base_admin.BaseCoreTest):
         new_name = 'Names'
         rename_mock = self.patch_call('cterasdk.core.files.io.rename')
         self.files.rename(path, new_name)
-        rename_mock.assert_called_once_with(self._global_admin, mock.ANY, new_name, wait=True)
+        rename_mock.assert_called_once_with(self._global_admin, mock.ANY, new_name)
         actual_ctera_path = rename_mock.call_args[0][1]
         self.assertEqual(actual_ctera_path.absolute, TestCoreFilesBrowser._create_expected_path(TestCoreFilesBrowser._base_path, path))
+        self._await_or_future_mock.assert_called_once_with(self._global_admin, mock.ANY, True)
 
     def test_delete(self):
         path = 'cloud/Users'
         rm_mock = self.patch_call('cterasdk.core.files.io.remove')
         self.files.delete(path)
-        rm_mock.assert_called_once_with(self._global_admin, mock.ANY, wait=True)
+        rm_mock.assert_called_once_with(self._global_admin, mock.ANY)
         actual_ctera_path = rm_mock.call_args[0][1]
         self.assertEqual(actual_ctera_path.absolute, TestCoreFilesBrowser._create_expected_path(TestCoreFilesBrowser._base_path, path))
+        self._await_or_future_mock.assert_called_once_with(self._global_admin, mock.ANY, True)
 
     def test_undelete(self):
         path = 'cloud/Users'
         recover_mock = self.patch_call('cterasdk.core.files.io.recover')
         self.files.undelete(path)
-        recover_mock.assert_called_once_with(self._global_admin, mock.ANY, wait=True)
+        recover_mock.assert_called_once_with(self._global_admin, mock.ANY)
         actual_ctera_path = recover_mock.call_args[0][1]
         self.assertEqual(actual_ctera_path.absolute, TestCoreFilesBrowser._create_expected_path(TestCoreFilesBrowser._base_path, path))
+        self._await_or_future_mock.assert_called_once_with(self._global_admin, mock.ANY, True)
 
     def test_move(self):
         src = 'cloud/Users'
         dst = 'public'
         mv_mock = self.patch_call('cterasdk.core.files.io.move')
         self.files.move(src, destination=dst)
-        mv_mock.assert_called_once_with(self._global_admin, mock.ANY, destination=mock.ANY, wait=True)
+        mv_mock.assert_called_once_with(self._global_admin, mock.ANY, destination=mock.ANY, resolver=None, cursor=None)
         actual_ctera_paths = mv_mock.call_args[0][1:]
         self.assertListEqual(
             [actual_ctera_path.absolute for actual_ctera_path in actual_ctera_paths],
             [TestCoreFilesBrowser._create_expected_path(TestCoreFilesBrowser._base_path, path) for path in [src]]
         )
+        self._await_or_future_mock.assert_called_once_with(self._global_admin, mock.ANY, True)
 
     def test_copy(self):
         src = 'cloud/Users'
         dst = 'public'
         cp_mock = self.patch_call('cterasdk.core.files.io.copy')
         self.files.copy(src, destination=dst)
-        cp_mock.assert_called_once_with(self._global_admin, mock.ANY, destination=mock.ANY, wait=True)
+        cp_mock.assert_called_once_with(self._global_admin, mock.ANY, destination=mock.ANY, resolver=None, cursor=None)
         actual_ctera_paths = cp_mock.call_args[0][1:]
         self.assertListEqual(
             [actual_ctera_path.absolute for actual_ctera_path in actual_ctera_paths],
             [TestCoreFilesBrowser._create_expected_path(TestCoreFilesBrowser._base_path, path) for path in [src]]
         )
+        self._await_or_future_mock.assert_called_once_with(self._global_admin, mock.ANY, True)
 
     def test_create_public_link_default_values(self):
         for access in [None, 'RW', 'RO']: